+++ predicators/ground_truth_models/__init__.py
@@ -135,13 +135,7 @@ utils.import_submodules(__path__, __name__)
 
 def deprecated_get_gt_nsrts(env_name: str) -> Set[NSRT]:
     """Create ground truth NSRTs for an env."""
-    if env_name == "cluttered_table":
-        nsrts = _get_cluttered_table_gt_nsrts(env_name)
-    elif env_name == "cluttered_table_place":
-        nsrts = _get_cluttered_table_gt_nsrts(env_name, with_place=True)
-    elif env_name in ("blocks", "pybullet_blocks", "blocks_clear"):
-        nsrts = _get_blocks_gt_nsrts(env_name)
-    elif env_name in ("painting", "repeated_nextto_painting"):
+    if env_name in ("painting", "repeated_nextto_painting"):
         nsrts = _get_painting_gt_nsrts(env_name)
     elif env_name == "tools":
         nsrts = _get_tools_gt_nsrts(env_name)
@@ -204,240 +198,6 @@ def _get_options_by_names(env_name: str,
     options = get_gt_options(env_name)
     name_to_option = {o.name: o for o in options}
     return [name_to_option[name] for name in names]
-
-
-def _get_cluttered_table_gt_nsrts(env_name: str,
-                                  with_place: bool = False) -> Set[NSRT]:
-    """Create ground truth NSRTs for ClutteredTableEnv."""
-    can_type, = _get_types_by_names(env_name, ["can"])
-
-    HandEmpty, Holding, Untrashed = _get_predicates_by_names(
-        env_name, ["HandEmpty", "Holding", "Untrashed"])
-
-    if with_place:
-        Grasp, Place = _get_options_by_names(env_name, ["Grasp", "Place"])
-    else:
-        Grasp, Dump = _get_options_by_names(env_name, ["Grasp", "Dump"])
-
-    nsrts = set()
-
-    # Grasp
-    can = Variable("?can", can_type)
-    parameters = [can]
-    option_vars = [can]
-    option = Grasp
-    preconditions = {LiftedAtom(HandEmpty, []), LiftedAtom(Untrashed, [can])}
-    add_effects = {LiftedAtom(Holding, [can])}
-    delete_effects = {LiftedAtom(HandEmpty, [])}
-
-    def grasp_sampler(state: State, goal: Set[GroundAtom],
-                      rng: np.random.Generator,
-                      objs: Sequence[Object]) -> Array:
-        del goal  # unused
-        assert len(objs) == 1
-        can = objs[0]
-        # Need a max here in case the can is trashed already, in which case
-        # both pose_x and pose_y will be -999.
-        end_x = max(0.0, state.get(can, "pose_x"))
-        end_y = max(0.0, state.get(can, "pose_y"))
-        if with_place:
-            start_x, start_y = 0.2, 0.1
-        else:
-            start_x, start_y = rng.uniform(0.0, 1.0,
-                                           size=2)  # start from anywhere
-        return np.array([start_x, start_y, end_x, end_y], dtype=np.float32)
-
-    grasp_nsrt = NSRT("Grasp",
-                      parameters, preconditions, add_effects, delete_effects,
-                      set(), option, option_vars, grasp_sampler)
-    nsrts.add(grasp_nsrt)
-
-    if not with_place:
-        # Dump
-        can = Variable("?can", can_type)
-        parameters = [can]
-        option_vars = []
-        option = Dump
-        preconditions = {
-            LiftedAtom(Holding, [can]),
-            LiftedAtom(Untrashed, [can])
-        }
-        add_effects = {LiftedAtom(HandEmpty, [])}
-        delete_effects = {
-            LiftedAtom(Holding, [can]),
-            LiftedAtom(Untrashed, [can])
-        }
-        dump_nsrt = NSRT("Dump", parameters, preconditions, add_effects,
-                         delete_effects, set(), option, option_vars,
-                         null_sampler)
-        nsrts.add(dump_nsrt)
-
-    else:
-        # Place
-        can = Variable("?can", can_type)
-        parameters = [can]
-        option_vars = [can]
-        option = Place
-        preconditions = {
-            LiftedAtom(Holding, [can]),
-            LiftedAtom(Untrashed, [can])
-        }
-        add_effects = {LiftedAtom(HandEmpty, [])}
-        delete_effects = {LiftedAtom(Holding, [can])}
-
-        def place_sampler(state: State, goal: Set[GroundAtom],
-                          rng: np.random.Generator,
-                          objs: Sequence[Object]) -> Array:
-            start_x, start_y = 0.2, 0.1
-            # Goal-conditioned sampling
-            if CFG.cluttered_table_place_goal_conditioned_sampling:
-                # Get the pose of the goal object
-                assert len(goal) == 1
-                goal_atom = next(iter(goal))
-                assert goal_atom.predicate == Holding
-                goal_obj = goal_atom.objects[0]
-                goal_x = state.get(goal_obj, "pose_x")
-                goal_y = state.get(goal_obj, "pose_y")
-                # Place up w.r.t the goal, and to some distance left
-                # or right such that we're not going out of x bounds
-                # 0 to 0.4.
-                end_y = goal_y * 1.1
-                end_x = goal_x + 0.2
-                if end_x > 0.4:
-                    end_x = goal_x - 0.2
-                return np.array([start_x, start_y, end_x, end_y],
-                                dtype=np.float32)
-            # Non-goal-conditioned sampling
-            del state, goal, objs
-            return np.array(
-                [start_x, start_y,
-                 rng.uniform(0, 0.4),
-                 rng.uniform(0, 1.0)],
-                dtype=np.float32)
-
-        place_nsrt = NSRT("Place", parameters, preconditions, add_effects,
-                          delete_effects, set(), option, option_vars,
-                          place_sampler)
-        nsrts.add(place_nsrt)
-
-    return nsrts
-
-
-def _get_blocks_gt_nsrts(env_name: str) -> Set[NSRT]:
-    """Create ground truth NSRTs for BlocksEnv."""
-    block_type, robot_type = _get_types_by_names(env_name, ["block", "robot"])
-
-    On, OnTable, GripperOpen, Holding, Clear = _get_predicates_by_names(
-        env_name, ["On", "OnTable", "GripperOpen", "Holding", "Clear"])
-
-    Pick, Stack, PutOnTable = _get_options_by_names(
-        env_name, ["Pick", "Stack", "PutOnTable"])
-
-    nsrts = set()
-
-    # PickFromTable
-    block = Variable("?block", block_type)
-    robot = Variable("?robot", robot_type)
-    parameters = [block, robot]
-    option_vars = [robot, block]
-    option = Pick
-    preconditions = {
-        LiftedAtom(OnTable, [block]),
-        LiftedAtom(Clear, [block]),
-        LiftedAtom(GripperOpen, [robot])
-    }
-    add_effects = {LiftedAtom(Holding, [block])}
-    delete_effects = {
-        LiftedAtom(OnTable, [block]),
-        LiftedAtom(Clear, [block]),
-        LiftedAtom(GripperOpen, [robot])
-    }
-
-    pickfromtable_nsrt = NSRT("PickFromTable", parameters, preconditions,
-                              add_effects, delete_effects, set(), option,
-                              option_vars, null_sampler)
-    nsrts.add(pickfromtable_nsrt)
-
-    # Unstack
-    block = Variable("?block", block_type)
-    otherblock = Variable("?otherblock", block_type)
-    robot = Variable("?robot", robot_type)
-    parameters = [block, otherblock, robot]
-    option_vars = [robot, block]
-    option = Pick
-    preconditions = {
-        LiftedAtom(On, [block, otherblock]),
-        LiftedAtom(Clear, [block]),
-        LiftedAtom(GripperOpen, [robot])
-    }
-    add_effects = {
-        LiftedAtom(Holding, [block]),
-        LiftedAtom(Clear, [otherblock])
-    }
-    delete_effects = {
-        LiftedAtom(On, [block, otherblock]),
-        LiftedAtom(Clear, [block]),
-        LiftedAtom(GripperOpen, [robot])
-    }
-    unstack_nsrt = NSRT("Unstack",
-                        parameters, preconditions, add_effects, delete_effects,
-                        set(), option, option_vars, null_sampler)
-    nsrts.add(unstack_nsrt)
-
-    # Stack
-    block = Variable("?block", block_type)
-    otherblock = Variable("?otherblock", block_type)
-    robot = Variable("?robot", robot_type)
-    parameters = [block, otherblock, robot]
-    option_vars = [robot, otherblock]
-    option = Stack
-    preconditions = {
-        LiftedAtom(Holding, [block]),
-        LiftedAtom(Clear, [otherblock])
-    }
-    add_effects = {
-        LiftedAtom(On, [block, otherblock]),
-        LiftedAtom(Clear, [block]),
-        LiftedAtom(GripperOpen, [robot])
-    }
-    delete_effects = {
-        LiftedAtom(Holding, [block]),
-        LiftedAtom(Clear, [otherblock])
-    }
-
-    stack_nsrt = NSRT("Stack", parameters, preconditions, add_effects,
-                      delete_effects, set(), option, option_vars, null_sampler)
-    nsrts.add(stack_nsrt)
-
-    # PutOnTable
-    block = Variable("?block", block_type)
-    robot = Variable("?robot", robot_type)
-    parameters = [block, robot]
-    option_vars = [robot]
-    option = PutOnTable
-    preconditions = {LiftedAtom(Holding, [block])}
-    add_effects = {
-        LiftedAtom(OnTable, [block]),
-        LiftedAtom(Clear, [block]),
-        LiftedAtom(GripperOpen, [robot])
-    }
-    delete_effects = {LiftedAtom(Holding, [block])}
-
-    def putontable_sampler(state: State, goal: Set[GroundAtom],
-                           rng: np.random.Generator,
-                           objs: Sequence[Object]) -> Array:
-        del state, goal, objs  # unused
-        # Note: normalized coordinates w.r.t. workspace.
-        x = rng.uniform()
-        y = rng.uniform()
-        return np.array([x, y], dtype=np.float32)
-
-    putontable_nsrt = NSRT("PutOnTable", parameters, preconditions,
-                           add_effects, delete_effects, set(), option,
-                           option_vars, putontable_sampler)
-    nsrts.add(putontable_nsrt)
-
-    return nsrts
 
 
 def _get_painting_gt_nsrts(env_name: str) -> Set[NSRT]:
