+++ predicators/ground_truth_models/cluttered_table/__init__.py
@@ -1,0 +1,5 @@
+"""Ground-truth models for cover environment and variants."""
+
+from .nsrts import ClutteredTableGroundTruthNSRTFactory
+
+__all__ = ["ClutteredTableGroundTruthNSRTFactory"]
